Bump .ruby-version to GitHub's Ruby 2.0.0.
I like vanilla Ruby as much as the next guy, but let's imitate the
infrastructure. Let's take the mission of imitating Pages' infrastructure to the
next level, by specifying that the Ruby version for development of this gem be
2.0.0-github. :sparkles:

For anyone who wishes to build GitHub's Ruby 2.0.0 with ruby-build, you can use
this:
https://github.com/parkr/rbenv-custom-definitions/blob/master/support/2.0.0-github

It's all about dat open-source.